function/thumbnail/aws_thumbnail/requirements.txt
Pillow
boto3
ntplib
smlibrary